Session Based Token Implementation
Tokens should now be session based instead of being stored in a local "token.json". The credential handling had to be reworked and used session middleware. Also, the credentials are retrieved in the path operation functions and passed to the functions they call instead of the functions they call retrieving the creds each time called - should reduce unnecessary overhead. Need to still implement the app opening the browser on startup. Does not automatically direct the user towards authenticating if unauthenticated.