Credential and exception related functions and refresh token fix
Added test_delete_creds(), exception_auth_redirect(), and improved error handling to redirect to /auth/login when refresh token is missing. Mostly ideal, except the application needs to be refreshed twice to bring up login button when refresh token is missing, first to delete the current credentials and next to bring up the button. To change dependent on creating an is_authenticated function which needs to be called by the frontend.